requirements.txt
newspaper3k
lxml_html_clean
requests-html
websockets==10.4
streamlit